src/ai_engine.py
import google.generativeai as genai
import json

genai.configure(api_key="YOUR_API_KEY")

model = genai.GenerativeModel("models/gemini-2.5-flash-lite")

SYSTEM_PROMPT = """
You are an AI product analyst.

The user input may include text and OCR extracted from images.
OCR text may contain spelling mistakes, broken grammar, or UI fragments.

You must infer the intended product even if the text is noisy or incomplete.

OCR from UI screenshots should be treated as a reference design, not as literal commands.

Extract:
- product_intent
- target_user
- core_features
- technical_constraints
- expected_outputs

Return only valid JSON.
If no reasonable product intent can be inferred, return:
{"status":"rejected","reason":"No product intent"}
"""

 
def analyze_input(text):
    prompt = f"""
You are an AI product analyst.

You must return ONLY valid JSON.
Do not include explanations or markdown.

OCR text may be noisy, misspelled, or incomplete.
You must infer the intended product.

Extract and return exactly this JSON structure:

{{
  "product_intent": "",
  "target_user": "",
  "core_features": [],
  "technical_constraints": [],
  "expected_outputs": []
}}

If no reasonable product intent can be inferred, return:
{{"status":"rejected","reason":"No product intent"}}

User input:
{text}
"""

    response = model.generate_content(prompt)

    raw = response.text.strip()

    # Extract JSON safely
    start = raw.find("{")
    end = raw.rfind("}") + 1
    clean_json = raw[start:end]

    return json.loads(clean_json)
