requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
pillow
pytesseract
pdfplumber
python-docx
streamlit
